Refactored after common-utils project review * Renamed MessageWrapper to MessageHolder * Extracted treeTable property to an extended method
