+++ src/main/kotlin/com/exactpro/th2/common/utils/message/MessageHolderUtils.kt
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2023 Exactpro (Exactpro Systems Limited)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.exactpro.th2.common.utils.message
+
+import com.exactpro.th2.common.event.bean.TreeTable
+import com.exactpro.th2.common.utils.message.transport.toTreeTable
+
+val MessageHolder.sessionAlias: String?
+    get() = id.sessionAlias
+
+fun MessageHolder.getTreeTable(): TreeTable = when (this) {
+    is ProtoMessageHolder -> source.toTreeTable()
+    is TransportMessageHolder -> source.toTreeTable()
+}
